1 lab algos finished

+++ AA/lab_01_new/src/alloc_mtrx.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void delete_mtrx(int **mtrx, size_t n)
+void delete_mtrx(int **mtrx, int n)
 {
     if (mtrx != NULL)
     {
-        for (size_t i = 0; i < n; i++)
+        for (int i = 0; i < n; i++)
         {
             if (mtrx[i] != NULL)
                 free(mtrx[i]);
@@ -15,14 +15,14 @@ void delete_mtrx(int **mtrx, size_t n)
     }
 }
 
-int **create_mtrx(size_t n, size_t m)
+int **create_mtrx(int n, int m)
 {
     if (n == 0 || m == 0)
         return NULL;
     int **mtrx = NULL;
     mtrx = (int**) malloc(sizeof(int *) * n);
 
-    for (size_t i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
         mtrx[i] = (int*) malloc(sizeof(int) * m);
     
     return mtrx;
